Code/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
html5lib>=1.1
openpyxl>=3.1.2
pytest>=8.0.0
